--- packages/web/config/ads-banner.ts
@@ -1,2 +1,0 @@
-export const ADS_BANNER_URL =
-  "https://raw.githubusercontent.com/osmosis-labs/fe-content/main/cms/swap-rotating-banner.json";
