+++ packages/web/config/index.ts
@@ -15,7 +15,5 @@ export * from "./feature-flag";
 export * from "./env";
 export * from "./ibc-overrides";
 export * from "./initial-assets-sort";
+export * from "./osmosis-repos";
 export * from "./user-analytics-v2";
-
-/** Config for displaying ads banner assets */
-export * from "./ads-banner";
