+++ packages/web/config/osmosis-repos.ts
@@ -1,0 +1,1 @@
+export const OsmosisCmsRepo = "osmosis-labs/fe-content";
